feat: separate rentals
