+++ packages/web/src/features/rental-business/components/Rentals/Easel.tsx
@@ -1,0 +1,8 @@
+import Typography from "@sparcs-clubs/web/common/components/Typography";
+import { RentalFrameProps } from "@sparcs-clubs/web/features/rental-business/frames/RentalNoticeFrame";
+
+const Easel: React.FC<RentalFrameProps> = () => (
+  <Typography type="p">대여 개수</Typography>
+);
+
+export default Easel;
