+++ packages/web/src/features/rental-business/components/Rentals/Vacuum.tsx
@@ -1,0 +1,24 @@
+import React from "react";
+import Radio from "@sparcs-clubs/web/common/components/Radio";
+import Typography from "@sparcs-clubs/web/common/components/Typography";
+import { RentalFrameProps } from "@sparcs-clubs/web/features/rental-business/frames/RentalNoticeFrame";
+
+const Vacuum: React.FC<RentalFrameProps> = ({ rental, setRental }) => (
+  <>
+    <Typography type="p">청소기 종류</Typography>
+    <Radio
+      value={rental?.vacuum || "cordless"}
+      onChange={value =>
+        setRental({
+          ...rental,
+          vacuum: value,
+        })
+      }
+    >
+      <Radio.Option value="cordless">무선 청소기</Radio.Option>
+      <Radio.Option value="corded">유선 청소기</Radio.Option>
+    </Radio>
+  </>
+);
+
+export default Vacuum;
